move module loading exception raising from deploy task to get_executor
