+++ src/inmanta/agent/executor.py
@@ -674,3 +674,75 @@ class ExecutorManager(abc.ABC, typing.Generic[E]):
         Any threadpools that need to be closed can be handed of to the parent via thread_pool_finalizer
         """
         pass
+
+
+class ModuleLoadingException(Exception):
+    """
+    This exception is raised when some Inmanta modules couldn't be loaded on a given agent.
+    """
+
+    def __init__(self, agent_name: str, failed_modules: FailedInmantaModules) -> None:
+        """
+        :param agent_name: Name of the agent for which module loading was unsuccessful
+        :param failed_modules: Data for all module loading errors as a nested map of
+            inmanta module name -> python module name -> Exception.
+        """
+        self.failed_modules = failed_modules
+        self.agent_name = agent_name
+
+    def _format_module_loading_errors(self) -> str:
+        """
+        Helper method to display module loading failures.
+        """
+        formatted_module_loading_errors = ""
+        N_FAILURES = sum(len(v) for v in self.failed_modules.values())
+        failure_index = 1
+
+        for _, failed_modules_data in self.failed_modules.items():
+            for python_module, exception in failed_modules_data.items():
+                formatted_module_loading_errors += f"Error {failure_index}/{N_FAILURES}:\n"
+                formatted_module_loading_errors += f"In module {python_module}:\n"
+                formatted_module_loading_errors += str(exception)
+                failure_index += 1
+
+        return formatted_module_loading_errors
+
+    def create_log_line_for_failed_modules(self, level: int = logging.ERROR, *, verbose_message: bool = False) -> LogLine:
+        """
+        Helper method to convert this Exception into a LogLine.
+
+        :param level: The log level for the resulting LogLine
+        :param verbose_message: Whether to include the full formatted error output in the LogLine message.
+            When displayed on the webconsole, the full formatted error output will be displayed in its own section
+            regardless of this flag's value.
+
+        """
+        message = "Agent %s failed to load the following modules: %s." % (
+            self.agent_name,
+            ", ".join(self.failed_modules.keys()),
+        )
+
+        formatted_module_loading_errors = self._format_module_loading_errors()
+
+        if verbose_message:
+            message += f"\n{formatted_module_loading_errors}"
+        return LogLine.log(
+            level=level,
+            msg=message,
+            timestamp=None,
+            errors=formatted_module_loading_errors,
+        )
+
+    def log_resource_action_to_scheduler_log(
+        self, agent: str, rid: ResourceVersionIdStr, *, include_exception_info: bool
+    ) -> None:
+        """
+        Helper method to log module loading failures to the scheduler's resource action log.
+
+        This method does not write anything to the 'resourceaction' table, which is what is ultimately displayed in the
+        'Logs' tab of the 'Resource Details' page in the web console. Therefore, the caller is responsible
+        for making sure that the scheduler's resource action log and the web console logs remain somewhat in sync.
+
+        """
+        log_line = self.create_log_line_for_failed_modules(verbose_message=True)
+        log_line.write_to_logger_for_resource(agent=agent, resource_version_string=rid, exc_info=include_exception_info)
